avgRating

+++ backend/db/migrations/20220914050931-create-review.js
@@ -32,7 +32,7 @@ module.exports = {
       },
       stars: {
         allowNull: true,
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL,
       },
       createdAt: {
         allowNull: false,
